+++ include/mruby_integration/core/input/mouse.hpp
@@ -1,6 +1,0 @@
-#pragma once
-
-#include "mruby.h"
-
-void
-append_core_input_mouse(mrb_state*);