+++ mrb_doc/core/input/keyboard.rb
@@ -1,8 +1,0 @@
-# Sets the key that {Window.close?} should listen for.
-# @param key [Integer]
-# @return [nil]
-def set_exit_key(key)
-  # mrb_set_exit_key
-  # src/mruby_integration/core/input/keyboard.cpp
-  nil
-end